Update flatbuffers-java instructions in README
Hello! First off love this repo, thanks for all the hard work :)

I attempted following the readme instructions and encountered a bunch of `NoSuchMethodError`s. It seems like this section of the README might be out of date, I've updated the POM example to the one that worked for me.

Hope this is correct, happy to make any necessary changes. Thanks!